+++ components/Form.tsx
@@ -1,5 +1,7 @@
-import React from "react";
-import countriesList from "utils/countriesList";
+import React, { useEffect, useState } from "react";
+import customFetch from "lib/customFetch";
+
+const APIURL = "/api/countrieslist";
 
 export default function Form({
   onChange,
@@ -12,6 +14,13 @@ export default function Form({
   method?: string;
   className?: string;
 }) {
+  const [countriesList, setCountriesList] = useState<string[]>([]);
+  useEffect(() => {
+    customFetch({ url: APIURL }).then(async (res) => {
+      const data = await res.json();
+      setCountriesList(data.countries);
+    });
+  }, []);
   return (
     <form
       onChange={onChange}
