+++ pages/api/data.ts
@@ -1,17 +1,9 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import fs from "fs";
-import path from "path";
 
 import _ from "lodash";
 import summaryComputingService from "services/summaryComputingService";
 import monthlyComputingService from "services/monthlyComputingService";
-import csvToArray from "utils/csvToArray";
-
-const FILE = "csv/memory-tech-challenge-data.csv";
-const file = fs.readFileSync(path.resolve("./public", FILE), {
-  encoding: "utf8",
-});
-const csv = csvToArray(file, /(?!\B"[^"]*),(?![^"]*"\B)/);
+import csv from "utils/csvParsing";
 
 //
 export default async function data(req: NextApiRequest, res: NextApiResponse) {
